+++ ecosysem/reactions.py
@@ -56,7 +56,7 @@ class KinP:
             Requested reaction.
         sample : STR or LIST, optional
             Requested samples (rows of `typeParam.csv`). The default is 'All'.
-        comp : STR or LIST, optional
+        compounds : STR or LIST, optional
             Compounds of parameters associated to compounds (e.g., Km). The default is None.
 
         Returns
@@ -67,6 +67,7 @@ class KinP:
             Names of samples (rows of `typeParam.csv`.
         """
         if not isinstance(params, list): params = [params]
+        if not isinstance(compounds, list): compounds = [compounds]
         dParam = pd.read_csv(KinP.path + paramDB + '.csv', encoding_errors='ignore')
         KinP.checkKinP(paramDB, dParam, params)
         dParam = dParam.loc[dParam['Reaction'] == reaction]
@@ -105,10 +106,10 @@ class KinRates:
 
         Parameters
         ----------
-        paramDB : STR
+        typeKin : STR
             Type of kinetic equations 
                 MM - 'Michaelis-Menten equation'.
-                MM-Arrhenihus - 'Michaelis-Menten-Arrhenius equation'
+                MM-Arrhenius - 'Michaelis-Menten-Arrhenius equation'
         paramDB : STR
             Name of parameter database, matching with csv name.
         reactions : STR
@@ -125,7 +126,7 @@ class KinRates:
         Returns
         -------
         Rs : DICT
-            Resultant rates. Shape: (Z)x(Y)x(X)x(reactions).
+            Resultant rates. Shape: (Z)x(Y)x(X).
         combNames : DICT
             Combination of samples (rows of `typeParam.csv`).
         orderComb : STR
@@ -158,6 +159,11 @@ class KinRates:
                 Ct[comp] = ThEq.pHSpeciation(comp, pH, T, Ct[comp])
         if typeKin == 'MM':
             params = ['qmax', 'Km']
+            """ # !!!
+            if len(paramDB) != 1:
+                print('!EcoSysEM.Error: only 1 database must be given: paramDB = "Michaelis-Menten DB"')
+                sys.exit()
+            """
             # Initialize results
             Rs = {}
             combNames = {}
@@ -253,7 +259,7 @@ class KinRates:
             k = Km[f'Km_{limComp}']
             if not isinstance(k, (list, np.ndarray)): k = np.array(k)
             if isinstance(k, np.ndarray):
-                if k.ndim != 1:
+                if k.ndim == 0:
                     k = np.array([k])
             K = k * np.ones(c.shape)
             M *= (c) / (K + c + 1e-25)
@@ -388,9 +394,10 @@ class Reactions:
         headers = np.empty(0)
         infoRxn = np.empty(0)
         for iCompound in compounds:
+            if iCompound == 'CO2': iCompound = 'H2CO3'
             dRxnAux1 = dRxn.filter(regex = f'^{iCompound}/')
             dRxnAux2 = dRxn.filter(regex = f'/{iCompound}/')
-            dRxnAux3 = dRxn.filter(regex = f'/{iCompound} *')
+            dRxnAux3 = dRxn.filter(regex = f'/{iCompound}\s')
             dRxnSizes = [dRxnAux1.shape[1], dRxnAux2.shape[1], dRxnAux3.shape[1]]
             ind_dRxnMax = dRxnSizes.index(max(dRxnSizes))
             if ind_dRxnMax == 0:
@@ -500,9 +507,9 @@ class Reactions:
             What reaction(s) type are requested, matching with csv name. E.g.:
                 - 'pHSpeciation': pH equilibrium
                 - 'metabolisms': metabolic activities
-        compounds : STR or LIST
-            Name(s) of requested compound(s).
-            STR - one compound; LIST - multiple compounds.
+        nameRxn : STR or LIST
+            Name(s) of requested reaction(s).
+            STR - one reaction; LIST - multiple reactions.
         warnings : BOOL, optional
             Display function warnings. The default is False.