+++ changelog.md
@@ -1,16 +1,32 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [Unreleased] - yyyy-mm-dd
+## [Unreleased] - yyyy/mm/dd
 ### Added
-- Sensitivity analysis of non-standard Gibbs free energy.
-  - 2D compound sensitivity analysis (contourf plot).
-  - Complete thermodynamic sensitivity analysis.
 - Bioenergetic calculations
   - Catabolic cell-specific power.
   - Empirical requirement cell-specific power.
 - New environment modelling framework.
   - Multi-Metabolic State Model (MMSM) based on cell-specific power.
+- Complete thermodynamic sensitivity analysis.
+### Changed
+- README file.
+
+## [0.3] - 2025/11/07
+### Added
+- New environment in `Hydrosphere`.
+  - General (or non-specific) Water Body: `GWB()`.
+- Thermodynamic State Analysis (ThSA) is incorporated in `Environment class`.
+  - ThSA functions are behaviours of `ISA()`, `ISAMERRA2()`, `CAMSMERRA2()` and `gWB()`.
+- Sensitivity analysis of non-standard Gibbs free energy.
+  - 2D compound sensitivity analysis (contourf plot).
+- Creation of plotting script
+  - New function: plotVarMap2D()
+  - New function: plotZonalMean()
+  - New function: plotCrossSections()
+### Fixed
+- Bug in `Environment.loadData()`: data existence wasn't check before loading.
+- Bug in `CAMSMERRA2()`: error came out when phase was defined.
 ### Changed
 - README file.